+++ src/main/java/net/silentchaos512/gear/gear/material/modifier/ChargedMaterialModifier.java
@@ -44,8 +44,10 @@ public abstract class ChargedMaterialModifier implements IMaterialModifier {
 
         @Override
         public void removeModifier(ItemStack stack) {
-            stack.getOrCreateTag().remove(nbtTagName);
-            stack.getOrCreateTag().remove(Const.NBT_IS_FOIL);
+            if (!stack.isEmpty()) {
+                stack.getOrCreateTag().remove(nbtTagName);
+                stack.getOrCreateTag().remove(Const.NBT_IS_FOIL);
+            }
         }
 
         boolean causesFoilEffect() {